requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.2.1
constantly==15.1.0
coverage==7.3.1
cryptography==41.0.4
cssselect==1.2.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.12.4
flake8==6.1.0
frida==16.1.4
frida-tools==12.3.0
frozenlist==1.4.0
greenlet==3.0.3
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.13.0
ipywidgets==8.1.1
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
json5==0.9.14
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter-server==2.12.5
jupyter-server-terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
jupyterlab-widgets==3.0.9
lxml==5.1.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
mock==5.1.0
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.0.7
notebook-shim==0.2.3
numpy==1.24.4
opencv-python==4.9.0.80
openpyxl==3.1.2
overrides==7.6.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
playwright==1.41.2
pluggy==1.3.0
prettytable==3.10.0
prometheus-client==0.19.0
prompt-toolkit==3.0.39
Protego==0.3.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pycryptodome==3.19.0
PyDispatcher==2.0.7
pyecharts==2.0.4
pyee==11.0.1
PyExecJS==1.5.1
pyflakes==3.1.0
Pygments==2.16.1
pyOpenSSL==23.2.0
pyppeteer==1.0.2
pyproject-api==1.6.1
pytest==7.4.2
pytest-base-url==2.1.0
pytest-cov==4.1.0
pytest-playwright==0.4.4
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
redis==5.0.0
referencing==0.32.1
requests==2.31.0
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
Scrapy==2.11.0
scrapy-redis==0.7.3
Send2Trash==1.8.2
service-identity==23.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
terminado==0.18.0
text-unidecode==1.3
tinycss2==1.2.1
tldextract==3.6.0
tomli==2.0.1
tornado==6.4
tox==4.11.3
tqdm==4.66.1
traitlets==5.14.1
Twisted==22.10.0
typing-extensions==4.8.0
tzdata==2023.4
urllib3==1.26.18
virtualenv==20.24.5
w3lib==2.1.2
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
widgetsnbextension==4.0.9
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
zipp==3.17.0
zope.interface==6.0
